+++ libraries/hostescalation.rb
@@ -29,21 +29,21 @@ class Nagios
                   :host_name,
                   :hostgroup_name,
                   :contacts,
-                  :contact_groups,
-                  :escalation_period
+                  :contact_groups
 
     attr_accessor :first_notification,
                   :last_notification,
                   :notification_interval,
-                  :escalation_options
+                  :escalation_options,
+                  :escalation_period
 
     def initialize(name)
-      @host_description, @name  = name
-      @contacts         = {}
-      @contact_groups   = {}
-      @host_name        = {}
-      @hostgroup_name   = {}
-      @register         = 0
+      @host_description, @name = name
+      @contacts       = {}
+      @contact_groups = {}
+      @host_name      = {}
+      @hostgroup_name = {}
+      @register       = 0
     end
 
     def definition