+++ recipes/pagerduty.rb
@@ -30,6 +30,30 @@ if using_old_pagerduty_key_attribute?
   node.set['nagios']['pagerduty']['key'] = node['nagios']['pagerduty_key']
 end
 
+package 'perl-CGI' do
+  case node['platform_family']
+  when 'rhel', 'fedora'
+    package_name 'perl-CGI'
+  when 'debian'
+    package_name 'libcgi-pm-perl'
+  when 'arch'
+    package_name 'perl-cgi'
+  end
+  action :install
+end
+
+package 'perl-JSON' do
+  case node['platform_family']
+  when 'rhel', 'fedora'
+    package_name 'perl-JSON'
+  when 'debian'
+    package_name 'libjson-perl'
+  when 'arch'
+    package_name 'perl-json'
+  end
+  action :install
+end
+
 package 'libwww-perl' do
   case node['platform_family']
   when 'rhel', 'fedora'
@@ -62,6 +86,16 @@ remote_file "#{node['nagios']['plugin_dir']}/notify_pagerduty.pl" do
   action :create_if_missing
 end
 
+template "#{node['nagios']['cgi-bin']}/pagerduty.cgi" do
+  source 'pagerduty.cgi.erb'
+  owner node['nagios']['user']
+  group node['nagios']['group']
+  mode '0755'
+  variables(
+    :command_file => node['nagios']['conf']['command_file']
+  )
+end
+
 nagios_bags = NagiosDataBags.new
 pagerduty_contacts = nagios_bags.get('nagios_pagerduty')
 
@@ -88,6 +122,7 @@ end
 
 pagerduty_contacts.each do |contact|
   name = contact['contact'] || contact['id']
+
   nagios_contact name do
     options 'alias'                         => "PagerDuty Pseudo-Contact #{name}",
             'service_notification_period'   => contact['service_notification_period'] || '24x7',
@@ -96,7 +131,8 @@ pagerduty_contacts.each do |contact|
             'host_notification_options'     => contact['host_notification_options'] || 'd,r',
             'service_notification_commands' => 'notify-service-by-pagerduty',
             'host_notification_commands'    => 'notify-host-by-pagerduty',
-            'pager'                         => contact['key'] || contact['pagerduty_key']
+            'pager'                         => contact['key'] || contact['pagerduty_key'],
+            'contactgroups'                 => contact['contactgroups']
   end
 end