+++ test/fixtures/cookbooks/nagios_test/metadata.rb
@@ -1,0 +1,7 @@
+name             'nagios_test'
+maintainer       'Sander Botman'
+maintainer_email 'sbotman@schubergphilis.com'
+license          'Apache 2.0'
+description      'Contains Nagios configuration items'
+long_description 'Creates configuration items for nagios in order for testing with test-kitchen.'
+version          '0.1.0'