+++ CHANGELOG.md
@@ -1,6 +1,59 @@
 nagios Cookbook CHANGELOG
 =========================
 This file is used to list changes made in each version of the nagios cookbook.
+7.2.4
+-----
+### Bug
+- #419 Fixing the nagios_interval logic and readme.
+- #421 Fixing loading of pagerduty databag contacts.
+- #430 Fixing loading of timeperiods out of databag with ducktyping.
+- #437 Fixing loading of unmanaged_host databag regards to environments.
+- #441 Enable setting of Fixnum's within nagios configuration attributes.
+
+### Improvement
+- #426 Added command: service_notify_by_sms_email.
+- #435 Adding pagerduty.cgi and needed packages
+
+7.2.2
+-----
+### Bug
+- Fixing the apache mpm breaking on centos.
+
+7.2.0
+-----
+### Testing
+- Added centos 7.1 for testing.
+- Added centos 5.11 for testing.
+- Added test-kitchen tests.
+
+### Improvement
+- Added logic to exclude nodes based on tag.
+- Including apache2::mpm_prefork for apache.
+- Added the ability to specify command arguments within services.
+- Added the ability to specify custom options on hosts, contacts and services.
+
+7.1.8
+-----
+### Bug
+- Fixing the unmanagedhosts databag filter on environment.
+- Fixing the services databag filter on environment.
+
+### Improvement
+- Moving the LWRP's providers into definitions.  
+  This will remove some extra complexity and output will be  
+  much nicer and debugging will be easier during the chef-converge.
+
+7.1.6
+-----
+### Bug
+- Fixing the nagios_[resource] provider delete action.
+
+### Improvement
+- Added option for custom apache auth based on attribute.
+- Update cgi-path attibute on source install.
+- Update on test-kitchen tests.
+- Update on kitchen-vagrant version.
+
 7.1.4
 -----
 ### Bug